+++ etc/check-package-size.sh
@@ -20,7 +20,7 @@ indent cargo diet -n --package-size-limit 25KB
 (enter git-ref && indent cargo diet -n --package-size-limit 4KB)
 (enter git-url && indent cargo diet -n --package-size-limit 6KB)
 (enter git-object && indent cargo diet -n --package-size-limit 15KB)
-(enter git-odb && indent cargo diet -n --package-size-limit 50KB)
+(enter git-odb && indent cargo diet -n --package-size-limit 55KB)
 (enter git-protocol && indent cargo diet -n --package-size-limit 20KB)
 (enter git-packetline && indent cargo diet -n --package-size-limit 7KB)
 (enter git-repository && indent cargo diet -n --package-size-limit 10KB)