+++ git-features/src/lib.rs
@@ -1,5 +1,6 @@
 #![forbid(unsafe_code, rust_2018_idioms)]
 
+pub mod fs;
 pub mod hash;
 pub mod interrupt;
 pub mod parallel;