+++ git-odb/src/compound/init.rs
@@ -1,22 +1,23 @@
 use crate::{compound, loose, pack};
-use quick_error::quick_error;
 use std::path::PathBuf;
 
-quick_error! {
-    #[derive(Debug)]
-    pub enum Error {
-        Pack(err: pack::bundle::Error) {
-            display("Failed to instantiate a pack bundle")
-            source(err)
-            from()
-        }
-    }
+#[derive(thiserror::Error, Debug)]
+pub enum Error {
+    #[error("The objects directory at '{0}' is not an accessible directory")]
+    Inaccessible(PathBuf),
+    #[error(transparent)]
+    Pack(#[from] pack::bundle::Error),
+    #[error(transparent)]
+    Alternate(#[from] Box<crate::alternate::Error>),
 }
 
 /// Instantiation
 impl compound::Db {
     pub fn at(objects_directory: impl Into<PathBuf>) -> Result<compound::Db, Error> {
         let loose_objects = objects_directory.into();
+        if !loose_objects.is_dir() {
+            return Err(Error::Inaccessible(loose_objects));
+        }
         let packs = if let Ok(entries) = std::fs::read_dir(loose_objects.join("packs")) {
             let mut packs_and_sizes = entries
                 .filter_map(Result::ok)
@@ -32,8 +33,9 @@ impl compound::Db {
         };
 
         Ok(compound::Db {
-            loose: loose::Db::at(loose_objects),
+            loose: loose::Db::at(loose_objects.clone()),
             packs,
+            alternates: crate::alternate::resolve(loose_objects).map_err(Box::new)?,
         })
     }
 }